+++ Nickel.InfoScreens/BasicInfoScreenRouteActionArgs.cs
@@ -1,0 +1,7 @@
+﻿namespace Nickel.InfoScreens;
+
+internal sealed class BasicInfoScreenRouteActionArgs : IInfoScreensApi.IBasicInfoScreenRoute.IAction.IArgs
+{
+	public G G { get; internal set; } = null!;
+	public IInfoScreensApi.IBasicInfoScreenRoute Route { get; internal set; } = null!;
+}