+++ Nickel.UpdateChecks.UI/ModEntry.cs
@@ -1,19 +1,15 @@
-﻿using HarmonyLib;
+﻿using FSPRO;
+using HarmonyLib;
 using Microsoft.Extensions.Logging;
-using Microsoft.Xna.Framework.Graphics;
 using Nanoray.Pintail;
 using Nanoray.PluginManager;
-using Nanoray.Shrike;
-using Nanoray.Shrike.Harmony;
 using Nickel.Common;
+using Nickel.InfoScreens;
 using Nickel.ModSettings;
 using System;
 using System.Collections.Generic;
 using System.Diagnostics;
-using System.Diagnostics.CodeAnalysis;
 using System.Linq;
-using System.Reflection;
-using System.Reflection.Emit;
 
 namespace Nickel.UpdateChecks.UI;
 
@@ -22,6 +18,7 @@ public sealed class ModEntry : SimpleMod
 	internal static ModEntry Instance { get; private set; } = null!;
 	internal readonly ILocaleBoundNonNullLocalizationProvider<IReadOnlyList<string>> Localizations;
 	internal readonly IUpdateChecksApi UpdateChecksApi;
+	internal readonly IModSettingsApi ModSettingsApi;
 
 	internal readonly Content Content;
 
@@ -37,6 +34,7 @@ public sealed class ModEntry : SimpleMod
 			)
 		);
 		this.UpdateChecksApi = helper.ModRegistry.GetApi<IUpdateChecksApi>("Nickel.UpdateChecks")!;
+		this.ModSettingsApi = helper.ModRegistry.GetApi<IModSettingsApi>("Nickel.ModSettings")!;
 		
 		this.Content = new()
 		{
@@ -48,20 +46,58 @@ public sealed class ModEntry : SimpleMod
 			DefaultVisitWebsiteOnIcon = this.Helper.Content.Sprites.RegisterSprite(this.Package.PackageRoot.GetRelativeFile("assets/DefaultVisitWebsiteOnIcon.png")),
 		};
 		
-		helper.ModRegistry.AwaitApi<IModSettingsApi>("Nickel.ModSettings", api =>
+		var updateChecksMod = helper.ModRegistry.LoadedMods.GetValueOrDefault("Nickel.UpdateChecks");
+		this.ModSettingsApi.OverrideModSettingsTitle(updateChecksMod?.DisplayName ?? updateChecksMod?.UniqueName);
+		this.ModSettingsApi.RegisterModSettings(this.MakeUpdateListModSetting());
+
+		if (helper.ModRegistry.GetApi<IInfoScreensApi>("Nickel.InfoScreens") is { } infoScreensApi)
 		{
-			var updateChecksMod = helper.ModRegistry.LoadedMods.GetValueOrDefault("Nickel.UpdateChecks");
-			api.OverrideModSettingsTitle(updateChecksMod?.DisplayName ?? updateChecksMod?.UniqueName);
-			api.RegisterModSettings(this.MakeUpdateListModSetting());
-		});
-		
-		var harmony = this.Helper.Utilities.Harmony;
-		
-		harmony.Patch(
-			original: AccessTools.DeclaredMethod(typeof(CornerMenu), nameof(CornerMenu.Render))
-			          ?? throw new InvalidOperationException($"Could not patch game methods: missing method `{nameof(CornerMenu)}.{nameof(CornerMenu.Render)}`"),
-			transpiler: new HarmonyMethod(MethodBase.GetCurrentMethod()!.DeclaringType!, nameof(CornerMenu_Render_Transpiler))
-		);
+			helper.Events.OnModLoadPhaseFinished += (_, phase) =>
+			{
+				if (phase != ModLoadPhase.AfterDbInit)
+					return;
+				
+				this.UpdateChecksApi.AwaitAllUpdateInfo(updates =>
+				{
+					var pendingUpdates = this.GetPendingModUpdates().ToList();
+					if (pendingUpdates.Count == 0)
+						return;
+				
+					var route = infoScreensApi.CreateBasicInfoScreenRoute();
+					route.Paragraphs = [
+						infoScreensApi.CreateBasicInfoScreenParagraph(this.Localizations.Localize(["infoScreen", "title"])).SetFont(DB.thicket),
+						infoScreensApi.CreateBasicInfoScreenParagraph(
+							pendingUpdates.Join(e => $"<c=textFaint>{this.UpdateChecksApi.GetModNameForUpdatePurposes(e.Mod)}</c> <c=textBold>{e.Mod.Version}</c> -> <c=boldPink>{e.Version}</c>", "\n")
+						),
+					];
+					route.Actions = [
+						infoScreensApi.CreateBasicInfoScreenAction(this.Localizations.Localize(["infoScreen", "actions", "details"]), args =>
+						{
+							Audio.Play(Event.Click);
+							args.Route.RouteOverride = this.ModSettingsApi.MakeModSettingsRouteForMod(package.Manifest);
+						}),
+						infoScreensApi.CreateBasicInfoScreenAction(this.Localizations.Localize(["infoScreen", "actions", "remindLater"]), args =>
+						{
+							Audio.Play(Event.Click);
+							args.G.CloseRoute(args.Route.AsRoute);
+						}).SetControllerKeybind(Btn.B),
+						infoScreensApi.CreateBasicInfoScreenAction(this.Localizations.Localize(["infoScreen", "actions", "ignore"]), args =>
+						{
+							Audio.Play(Event.Click);
+							
+							foreach (var kvp in updates)
+								if (kvp.Value is not null && kvp.Value.Count != 0)
+									this.UpdateChecksApi.SetIgnoredUpdateForMod(kvp.Key, kvp.Value.Select(u => u.Version).Max());
+							this.UpdateChecksApi.SaveSettings();
+							
+							args.G.CloseRoute(args.Route.AsRoute);
+						}).SetColor(Colors.redd).SetRequiresConfirmation(true),
+					];
+
+					infoScreensApi.RequestInfoScreen("UpdatesAvailable", route.AsRoute, 1_000);
+				});
+			};
+		}
 	}
 
 	public override object GetApi(IModManifest requestingMod)
@@ -129,143 +165,5 @@ public sealed class ModEntry : SimpleMod
 		).SubscribeToOnMenuClose(
 			_ => this.UpdateChecksApi.SaveSettings()
 		));
-	}
-
-	[SuppressMessage("ReSharper", "PossibleMultipleEnumeration")]
-	private static IEnumerable<CodeInstruction> CornerMenu_Render_Transpiler(IEnumerable<CodeInstruction> instructions, MethodBase originalMethod)
-	{
-		try
-		{
-			return new SequenceBlockMatcher<CodeInstruction>(instructions)
-				.Find(ILMatches.LdcI4((int)StableSpr.buttons_menu))
-				.Find(ILMatches.Call("Sprite"))
-				.Replace(new CodeInstruction(OpCodes.Call, AccessTools.DeclaredMethod(MethodBase.GetCurrentMethod()!.DeclaringType!, nameof(CornerMenu_Render_Transpiler_HijackDraw))))
-				.AllElements();
-		}
-		catch (Exception ex)
-		{
-			Instance.Logger.LogError("Could not patch method {Method} - {Mod} probably won't work.\nReason: {Exception}", originalMethod, Instance.Package.Manifest.UniqueName, ex);
-			return instructions;
-		}
-	}
-
-	private static void CornerMenu_Render_Transpiler_HijackDraw(Spr? id, double x, double y, bool flipX, bool flipY, double rotation, Vec? originPx, Vec? originRel, Vec? scale, Rect? pixelRect, Color? color, BlendState? blend, SamplerState? samplerState, Effect? effect)
-	{
-		Draw.Sprite(id, x, y, flipX, flipY, rotation, originPx, originRel, scale, pixelRect, color, blend, samplerState, effect);
-
-		var updateSourceMessages = Instance.UpdateChecksApi.UpdateSources
-			.OrderBy(kvp => kvp.Key)
-			.Select(kvp => kvp.Value)
-			.Select(s => Instance.AsUiUpdateSource(s))
-			.OfType<IUiUpdateSource>()
-			.SelectMany(s => s.Messages)
-			.GroupBy(m => m.Level)
-			.ToDictionary(g => g.Key, g => g.ToList());
-
-		var updatesAvailable = Instance.GetPendingModUpdates()
-			.Where(e => Instance.UpdateChecksApi.GetIgnoredUpdateForMod(e.Mod) != e.Version)
-			.ToList();
-
-		List<ISpriteEntry> overlaysToShow = [];
-		var addedTooltips = false;
-
-		if (updatesAvailable.Count > 0)
-			overlaysToShow.Add(Instance.Content.UpdateAvailableOverlayIcon);
-
-		if (updateSourceMessages.TryGetValue(UpdateSourceMessageLevel.Error, out var messages) && messages.Count > 0)
-			overlaysToShow.Add(Instance.Content.ErrorMessageOverlayIcon);
-		else if (updateSourceMessages.TryGetValue(UpdateSourceMessageLevel.Warning, out messages) && messages.Count > 0)
-			overlaysToShow.Add(Instance.Content.WarningMessageOverlayIcon);
-
-		if (overlaysToShow.Count > 0)
-		{
-			var overlayToShow = overlaysToShow[(int)MG.inst.g.time % overlaysToShow.Count];
-			Draw.Sprite(overlayToShow.Sprite, x, y);
-		}
-
-		if (MG.inst.g.boxes.FirstOrDefault(b => b.key is { } key && key.k == StableUK.corner_mainmenu) is not { } box)
-			return;
-		if (!box.IsHover())
-			return;
-
-		if (updateSourceMessages.TryGetValue(UpdateSourceMessageLevel.Error, out messages) && messages.Count > 0)
-		{
-			if (!addedTooltips)
-			{
-				addedTooltips = true;
-				MG.inst.g.tooltips.Add(box.rect.xy + new Vec(15, 15), new TTDivider());
-			}
-
-			var i = 0;
-			foreach (var error in messages)
-			{
-				MG.inst.g.tooltips.Add(box.rect.xy + new Vec(15, 15), new GlossaryTooltip($"ui.{Instance.Package.Manifest.UniqueName}::Error{i++}")
-				{
-					Icon = StableSpr.icons_hurt,
-					TitleColor = Colors.textBold,
-					Title = Instance.Localizations.Localize(["settingsTooltip", "error"]),
-					Description = error.Message,
-				});
-			}
-		}
-
-		if (updateSourceMessages.TryGetValue(UpdateSourceMessageLevel.Warning, out messages) && messages.Count > 0)
-		{
-			if (!addedTooltips)
-			{
-				addedTooltips = true;
-				MG.inst.g.tooltips.Add(box.rect.xy + new Vec(15, 15), new TTDivider());
-			}
-
-			var i = 0;
-			foreach (var error in messages)
-			{
-				MG.inst.g.tooltips.Add(box.rect.xy + new Vec(15, 15), new GlossaryTooltip($"ui.{Instance.Package.Manifest.UniqueName}::Warning{i++}")
-				{
-					Icon = StableSpr.icons_hurtBlockable,
-					TitleColor = Colors.textBold,
-					Title = Instance.Localizations.Localize(["settingsTooltip", "warning"]),
-					Description = error.Message,
-				});
-			}
-		}
-
-		if (updateSourceMessages.TryGetValue(UpdateSourceMessageLevel.Info, out messages) && messages.Count > 0)
-		{
-			if (!addedTooltips)
-			{
-				addedTooltips = true;
-				MG.inst.g.tooltips.Add(box.rect.xy + new Vec(15, 15), new TTDivider());
-			}
-
-			var i = 0;
-			foreach (var error in messages)
-			{
-				MG.inst.g.tooltips.Add(box.rect.xy + new Vec(15, 15), new GlossaryTooltip($"ui.{Instance.Package.Manifest.UniqueName}::Info{i++}")
-				{
-					Icon = StableSpr.icons_hurtBlockable,
-					TitleColor = Colors.textBold,
-					Title = Instance.Localizations.Localize(["settingsTooltip", "info"]),
-					Description = error.Message,
-				});
-			}
-		}
-
-		if (updatesAvailable.Count > 0)
-		{
-			if (!addedTooltips)
-			{
-				// addedTooltips = true;
-				MG.inst.g.tooltips.Add(box.rect.xy + new Vec(15, 15), new TTDivider());
-			}
-
-			MG.inst.g.tooltips.Add(box.rect.xy + new Vec(15, 15), new GlossaryTooltip($"ui.{Instance.Package.Manifest.UniqueName}::UpdatesAvailable")
-			{
-				Icon = Instance.Content.UpdateAvailableTooltipIcon.Sprite,
-				TitleColor = Colors.textBold,
-				Title = Instance.Localizations.Localize(["settingsTooltip", "updatesAvailableTooltip"]),
-				Description = string.Join("\n", updatesAvailable.Select(e => $"<c=textFaint>{Instance.UpdateChecksApi.GetModNameForUpdatePurposes(e.Mod)}</c> <c=textBold>{e.Mod.Version}</c> -> <c=boldPink>{e.Version}</c>"))
-			});
-		}
 	}
 }